requirements.txt
agate==1.6.3
alembic==1.7.7
anyio==3.5.0
apache-airflow==2.2.5
apache-airflow-providers-airbyte==2.1.4
apache-airflow-providers-ftp==2.1.2
apache-airflow-providers-google==6.8.0
apache-airflow-providers-http==2.1.2
apache-airflow-providers-imap==2.2.3
apache-airflow-providers-postgres==4.1.0
apache-airflow-providers-sqlite==2.1.3
apispec==3.3.2
argcomplete==1.12.3
attrs==20.3.0
Babel==2.10.1
blinker==1.4
cachelib==0.6.0
cachetools==5.0.0
cattrs==1.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==7.1.2
clickclick==20.10.2
colorama==0.4.4
colorlog==4.8.0
commonmark==0.9.1
connexion==2.13.1
croniter==1.3.4
cryptography==37.0.1
db-dtypes==1.0.0
dbt-bigquery==1.1.0
dbt-core==1.1.0
dbt-extractor==0.4.1
dbt-postgres==1.1.0
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.4
distlib==0.3.4
dnspython==2.2.1
docutils==0.16
email-validator==1.2.1
fastapi==0.75.2
filelock==3.6.0
Flask==1.1.4
Flask-AppBuilder==3.4.5
Flask-Babel==2.0.0
Flask-Caching==1.10.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
future==0.18.2
google-ads==14.0.0
google-api-core==2.7.3
google-api-python-client==1.12.11
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
google-cloud-aiplatform==1.12.1
google-cloud-appengine-logging==1.1.1
google-cloud-audit-log==0.2.0
google-cloud-automl==2.7.2
google-cloud-bigquery==2.34.3
google-cloud-bigquery-datatransfer==3.6.1
google-cloud-bigquery-storage==2.13.1
google-cloud-bigtable==1.7.1
google-cloud-build==3.8.1
google-cloud-container==1.0.2
google-cloud-core==2.3.0
google-cloud-datacatalog==3.7.1
google-cloud-dataplex==1.0.0
google-cloud-dataproc==4.0.2
google-cloud-dataproc-metastore==1.5.0
google-cloud-dlp==1.0.1
google-cloud-kms==2.11.1
google-cloud-language==1.3.1
google-cloud-logging==3.0.0
google-cloud-memcache==1.3.1
google-cloud-monitoring==2.9.1
google-cloud-orchestration-airflow==1.3.1
google-cloud-os-login==2.6.1
google-cloud-pubsub==2.12.0
google-cloud-redis==2.8.0
google-cloud-resource-manager==1.4.1
google-cloud-secret-manager==1.0.1
google-cloud-spanner==1.19.2
google-cloud-speech==1.3.3
google-cloud-storage==1.44.0
google-cloud-tasks==2.8.1
google-cloud-texttospeech==1.0.2
google-cloud-translate==1.7.1
google-cloud-videointelligence==1.16.2
google-cloud-vision==1.0.1
google-cloud-workflows==1.6.1
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
graphviz==0.20
grpc-google-iam-v1==0.12.4
grpcio==1.44.0
grpcio-gcp==0.2.2
grpcio-status==1.44.0
gunicorn==20.1.0
h11==0.12.0
hologram==0.0.14
httpcore==0.14.7
httplib2==0.20.4
httpx==0.22.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
inflection==0.5.1
iniconfig==1.1.1
iso8601==1.0.2
isodate==0.6.1
itsdangerous==1.1.0
Jinja2==2.11.3
json-merge-patch==0.2
jsonschema==3.1.1
lazy-object-proxy==1.7.1
leather==0.3.4
lockfile==0.12.2
Logbook==1.5.3
looker-sdk==22.4.0
Mako==1.2.0
Markdown==3.3.6
MarkupSafe==2.0.1
marshmallow==3.15.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mashumaro==2.9
minimal-snowplow-tracker==0.0.2
msgpack==1.0.3
networkx==2.8
nox==2020.12.31
numpy==1.22.3
oauthlib==3.2.0
packaging==21.3
pandas==1.4.2
pandas-gbq==0.17.4
parsedatetime==2.4
pendulum==2.1.2
platformdirs==2.5.2
pluggy==1.0.0
prison==0.2.1
proto-plus==1.18.1
protobuf==3.20.1
psutil==5.9.0
psycopg2-binary==2.9.3
py==1.11.0
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.0
pydata-google-auth==1.4.0
Pygments==2.12.0
PyJWT==1.7.1
pyOpenSSL==22.0.0
pyparsing==3.0.8
pyrsistent==0.18.1
pytest==7.1.2
python-daemon==2.3.0
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytimeparse==1.1.8
pytz==2022.1
pytzdata==2020.1
PyYAML==5.4.1
requests==2.27.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==12.3.0
rsa==4.8
setproctitle==1.2.3
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.3.24
sqlalchemy-bigquery==1.4.3
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.38.2
sqlparse==0.4.2
starlette==0.17.1
swagger-ui-bundle==0.0.9
tabulate==0.8.9
tenacity==8.0.1
termcolor==1.1.0
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.2.0
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.26.9
virtualenv==20.14.1
Werkzeug==1.0.1
wrapt==1.14.1
WTForms==2.3.3
zipp==3.8.0
